move spectral utils to new module
