+++ mikeio/dfsu_spectral.py
@@ -9,8 +9,7 @@ from mikecore.DfsuFile import DfsuFile, DfsuFileType
 from .dfsu import _Dfsu
 from .dataset import Dataset, DataArray
 from .dfsutil import _get_item_info, _valid_item_numbers, _valid_timesteps
-from .spatial.FM_utils import _plot_spectrum, _calc_m0_from_spectrum
-
+from .spectral_utils import _plot_spectrum, _calc_m0_from_spectrum
 
 class DfsuSpectral(_Dfsu):
     @property
